Bring nested UnderFileStatus into its own class
