+++ core/common/src/main/java/alluxio/underfs/UnderFileSystem.java
@@ -189,57 +189,6 @@ public abstract class UnderFileSystem {
   }
 
   /**
-   * Information about a file or a directory in the under file system.
-   */
-  protected static class UnderFileStatus {
-    private final boolean mIsDirectory;
-    private final String mName;
-
-    /**
-     * Create new instance for under file information.
-     *
-     * @param name relative path of file or directory
-     * @param isDirectory whether the path is a directory
-     */
-    public UnderFileStatus(String name, boolean isDirectory) {
-      mIsDirectory = isDirectory;
-      mName = name;
-    }
-
-    public boolean isDirectory() {
-      return mIsDirectory;
-    }
-
-    public boolean isFile() {
-      return !mIsDirectory;
-    }
-
-    public String getName() {
-      return mName;
-    }
-
-    @Override
-    public String toString() {
-      return getName();
-    }
-
-    /**
-     * Convert an array of UFS file status to a listing result where each element in the array is
-     * a file or directory name.
-     *
-     * @param children array of internal listing
-     * @return array of file or directory names
-     */
-    public static String[] toListingResult(UnderFileStatus[] children) {
-      String[] ret = new String[children.length];
-      for (int i = 0; i < children.length; ++i) {
-        ret[i] = children[i].getName();
-      }
-      return ret;
-    }
-  }
-
-  /**
    * Gets the UnderFileSystem instance according to its schema.
    *
    * @param path the file path storing over the ufs
